+++ src/hypernix/__init__.py
@@ -242,5 +242,5 @@ __all__ = [
     "whisk",
 ]
 
-__version__ = "0.61.1"
+__version__ = "0.61.2"
 DEFAULT_REPO_ID = "ray0rf1re/hyper-Nix.2"